+++ CHANGELOG.md
@@ -8,6 +8,7 @@ ADDED:
 
 - new decorator method in class ThreadExtensions for get the available processors (cores) on the current machine
 - new decorator method in class ThreadExtensions for get the half of the available processors (cores) on the current machine
+- new method in class ThreadExtensions for shut down the given executor service with a given time out
 
 Version 8.3
 -------------